code refactor

+++ src/canvass/canvass.service.ts
@@ -1,18 +1,19 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { CreateCanvassInput } from './dto/create-canvass.input';
 import { UpdateCanvassInput } from './dto/update-canvass.input';
 import { PrismaService } from 'src/__prisma__/prisma.service';
 import { Canvass } from '@prisma/client';
-import { CommonService } from 'src/__common__/common.service';
+import { CommonPurchasingService } from 'src/__common__/common.purchasing.service';
+import { NotFoundError } from 'rxjs';
 
 @Injectable()
 export class CanvassService {
 
-  private readonly logger = new Logger(CommonService.name);
+  private readonly logger = new Logger(CommonPurchasingService.name);
 
   constructor(
     private readonly prisma: PrismaService,
-    private readonly common: CommonService
+    private readonly purchasing: CommonPurchasingService
   ){}
 
   // prisma behind the scenes uses db transaction
@@ -21,18 +22,11 @@ export class CanvassService {
 
     try {
 
-      await this.common.validateRcNumberUnique({
+      await this.purchasing.validateRcNumberUnique({
           rcNumber: input.rc_number,
           table: 'canvass',
           field: 'rc_number'
       })
-
-      await this.common.validateIds({
-        requested_by_id: input.requested_by_id,
-        noted_by_id: input.noted_by_id,
-      });
-
-      await this.common.validateItems(input.items)
 
       const createdCanvass = await this.prisma.canvass.create({
         data: {
@@ -56,12 +50,18 @@ export class CanvassService {
           },
         },
       });
-    
-      return createdCanvass;
+      
+      this.logger.log('Successfully created canvass')
+      return await this.findOne(createdCanvass.id)
       
     } catch (error) {
       this.logger.error('Error creating canvass:', error);
-      throw new Error('Could not create canvass. Please try again.'); 
+
+      if(error instanceof BadRequestException || error instanceof ConflictException) {
+        throw error
+      }
+
+      throw new Error('Could not create canvass. Please refresh the page and try again.'); 
     }
   
   }
@@ -84,6 +84,9 @@ export class CanvassService {
       },
       where: {
         is_deleted: false
+      },
+      orderBy: {
+        rc_number: 'desc'
       }
     });
   }
@@ -116,10 +119,14 @@ export class CanvassService {
     const transactionResult = await this.prisma.$executeRaw`BEGIN`;
     
     try {
-      const existingCanvass = await this.prisma.canvass.findUniqueOrThrow({
+      const existingCanvass = await this.prisma.canvass.findUnique({
         where: { id },
         include: { canvass_items: { include: { item: true } } },
       });
+
+      if(!existingCanvass){
+        throw new NotFoundException(`Canvass with ID ${id} not found`)
+      }
 
       // Delete existing canvass_items
       await this.prisma.canvassItem.deleteMany({
@@ -160,14 +167,18 @@ export class CanvassService {
 
       await this.prisma.$executeRaw`COMMIT`;
       
-      return updatedCanvass;
+      return await this.findOne(updatedCanvass.id)
     } catch (error) {
       this.logger.error('Error updating canvass:', error);
       
       // Rollback the transaction on error
       await this.prisma.$executeRaw`ROLLBACK`;
 
-      throw new Error(error);
+      if(error instanceof BadRequestException || error instanceof ConflictException || error instanceof NotFoundException) {
+        throw error
+      }
+
+      throw new Error('Could not create canvass. Please refresh the page and try again.'); 
     }
   }
 
@@ -176,10 +187,6 @@ export class CanvassService {
     const item = await this.prisma.canvass.findUniqueOrThrow({
         where: {id, is_deleted: false}
     })
-
-    if(!item){
-      throw NotFoundException
-    }
     
     await this.prisma.canvass.update({
         where: { id },
@@ -194,7 +201,7 @@ export class CanvassService {
 
   async findLatestRcNumber(): Promise<string> {
     try {
-        return await this.common.getLatestRcNumber({
+        return await this.purchasing.getLatestRcNumber({
           table: 'canvass',
           field: 'rc_number'
         })
